% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lookup.R
\name{create_lookup}
\alias{create_lookup}
\title{Create a dictionary of terms to lookup in document-feature matrices}
\usage{
create_lookup(...)
}
\arguments{
\item{...}{}
}
\value{
A dictionary of terms
}
\description{
Create a dictionary of terms to lookup in document-feature matrices
}
\examples{
create_lookup(cars = c("Ford", "Mazda", "Lotus"))
}
