% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dsph_list.R
\name{get_dsph_england}
\alias{get_dsph_england}
\title{Get a list of Directors of Public Health in England}
\usage{
get_dsph_england(
  url =
    "https://www.gov.uk/government/publications/directors-of-public-health-in-england--2/directors-of-public-health-in-england"
)
}
\arguments{
\item{url}{}
}
\value{
A dataframe of local authority districts and their current Director of Public Health
}
\description{
Get a list of Directors of Public Health in England
}
\examples{

dsph <- get_dsph_england()

}
