% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_page_csv.R
\name{get_page_csvs}
\alias{get_page_csvs}
\title{Find csv/xls(x) file links on a web page}
\usage{
get_page_csvs(url)
}
\arguments{
\item{url}{}
}
\value{
A list of links to csv/ xls(x) files
}
\description{
Find csv/xls(x) file links on a web page
}
\examples{
url <- "https://fingertips.phe.org.uk"
csv <- get_page_csvs(url)
}
