% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_phe_catalogue.R
\name{get_phe_catalogue}
\alias{get_phe_catalogue}
\title{Get an interactive datatable of PHE resources on GOV.UK}
\usage{
get_phe_catalogue(
  url = "https://www.gov.uk/search/all?organisations\%5B\%5D=public-health-england",
  n = 298
)
}
\arguments{
\item{url}{}

\item{n}{}

\item{pages}{}
}
\value{
Returns a datatable of PHE resources
}
\description{
Get an interactive datatable of PHE resources on GOV.UK
}
\examples{
df <- get_phe_catalogue(n = 100)
}
