% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_viz.R
\name{page_viz}
\alias{page_viz}
\title{Visualise text pages with keywords highlighted as a page plot}
\usage{
page_viz(text, keyword, title = "Title")
}
\arguments{
\item{text}{}

\item{keyword}{}

\item{title}{}
}
\value{
A page visualisation where selected keywords are highlighted
}
\description{
Visualise text pages with keywords highlighted as a page plot
}
\examples{

page_viz(text, keyword = "fingertips", title = "Mentions of Fingertips in document")

}
